5001/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy-stl==3.1.1
trimesh==4.0.10
lib3mf==2.3.1
pillow==10.1.0
numpy==1.26.2
networkx==3.2.1
gunicorn==21.2.0
python-dotenv==1.0.0
PyShopify==0.9.11
requests==2.31.0

